requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.1
pillow>=10.3.0
httpx>=0.27.0

